general cleanup

+++ crates/inel/tests/unit/io/read.rs
@@ -119,7 +119,7 @@ fn cancel() {
     std::fs::write(&name, [b'a'; 128_000]).unwrap();
 
     for _ in 0..1000 {
-        let index = rand::random::<usize>() % 16;
+        let index = rand::random::<u64>() as usize % 16;
         let name_clone = name.clone();
 
         inel::block_on(async move {
@@ -146,7 +146,7 @@ fn cancel() {
     }
 
     for _ in 0..1000 {
-        let index = rand::random::<usize>() % 16;
+        let index = rand::random::<u64>() as usize % 16;
         let name_clone = name.clone();
 
         inel::block_on(async move {
